Number of templates allowed is 40. Handling and receiving errors

+++ server/src/routes/templates.routes.js
@@ -17,12 +17,18 @@ router.post('/templates', isAuthenticated, async (req, res) => {
   const data = body.map(template => ([template.title, template.position, user]))
 
   try {
+
+    if (data.length > 40) throw new Error(`Max template limit exceeded. Max number of templates: 40 `)
+
     await pool.query(`DELETE FROM templates WHERE user=?`, [req.user])
     await pool.query(`INSERT INTO templates(title, position, user) VALUES ?`, [data])
     const [result] = await pool.query(`SELECT * FROM templates WHERE user=?`, req.user)
     res.json(result)
   } catch (err) {
-    console.log({ err })
+    res.json({
+      error: true,
+      message: err.message
+    })
   }
 })
